requirements.txt
asgiref==3.5.2
awsebcli==3.20.3
botocore==1.23.54
cement==2.8.2
certifi==2022.9.24
charset-normalizer==2.0.12
colorama==0.4.3
cx-Oracle==8.3.0
Django==3.2.16
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
future==0.16.0
idna==3.4
jmespath==0.10.0
pathspec==0.9.0
pypiwin32==223
python-dateutil==2.8.2
pytz==2022.6
pywin32==305
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
termcolor==1.1.0
typing_extensions==4.4.0
urllib3==1.26.12
wcwidth==0.1.9
